Mention client socket timeout setting
A common configuration for DavMail is a server running with one long-running
client (e.g. mutt) connecting via localhost. The default configuration of
DavMail will cause the client to lose connection after a relatively short
period of inactivity, with nothing in the server log unless logging is
cranked up to DEBUG.