IMAP: fix: quoted special backslash in passwords
This change introduces a method StringUtil::parseQuotedImapString that
will return the string value of a quoted-string according to the
RFC 3501 (imap v4rev1) ABNF. While the method might not be concise, it
hopefully is correct.

This added method is then used in ImapConnection to correctly unquote
a quoted string token, which was previously done with some deviations
from RFC3501. Specifically, when the quoted string token was a password
transporting a string value that (also) contained one backslash (encoded
as a double-backslash), it would not be unquoted correctly to a single
backslash.

As a side-effect, ImapConnection::parseCredentials() needs not unquote
the first backslash in a username anymore, as all backslashes in all
quoted strings are unquoted correctly.

In order to limit the change in size, the previously existing processing
in ImapConnection::ImapTokenizer::nextToken() is left in-tact by still
calling removeQuotes() in all other cases.

Patch By: Max-Julian Pogner <max-julian@pogner.at>