IMAP: simplify implementation of keep-alive in loadFolder()
Perform the long-running in the main thread, and start a parallel thread
to send the keep-alive message at regular interval.  This makes the code
more regular and the feature more encapsulated.  The parallel thread is
started only when the number of messages to load is greater than 500,
and the keep-alive feature is enables.  The keep-alive thread is
canceled as soon as the folder has loaded its messages.  The keep-alive
task is scheduled to be sent every 20 s, after an initial delay of 20 s.